Move hooks constants to types and changed the imports
