+++ src/components/Garden/ActionCollateral.tsx
@@ -4,7 +4,7 @@ import { GU } from '@1hive/1hive-ui'
 import useGardenTokenIcon from '@hooks/useGardenTokenIcon'
 import { formatTokenAmount } from '@utils/token-utils'
 import lockIconSvg from '@assets/icon-lock.svg'
-import { ProposalType } from '@/hooks/constants'
+import { ProposalType } from '@/types/app'
 
 type ActionCollateralProps = {
   proposal: ProposalType
