+++ src/components/Garden/Feed/Delegation.tsx
@@ -13,6 +13,8 @@ import IdentityBadge from '@components/IdentityBadge'
 import useProfile from '@hooks/useProfile'
 import { useSupporterSubscription } from '@hooks/useSubscriptions'
 import { useWallet } from '@providers/Wallet'
+
+//TODO: This import will be replaced with @1hive/1hive-ui after new types is updated
 import { HiveUiElements } from './Icons'
 
 type DelegationProps = {
