+++ src/components/Garden/Feed/ProposalsList.tsx
@@ -21,7 +21,7 @@ import ProposalCard from './ProposalCard'
 import ProposalRankings from './ProposalRankings'
 
 import filterToggleSvg from '@assets/filter.svg'
-import { ProposalType } from '@/hooks/constants'
+import { ProposalType } from '@/types/app'
 import AbstainCard from './AbstainCard'
 import { ABSTAIN_PROPOSAL } from '../ProposalDetail/ProposalStatus'
