+++ src/components/Garden/ProposalDetail/ProposalHeader.tsx
@@ -3,7 +3,7 @@ import { GU, textStyle, useTheme } from '@1hive/1hive-ui'
 import ProposalIcon from '@components/ProposalIcon'
 import { dateFormat } from '@utils/date-utils'
 import { convertToString } from '@/types'
-import { ProposalType } from '@/hooks/constants'
+import { ProposalType } from '@/types/app'
 
 type ProposalHeaderProps = {
   proposal: ProposalType
