+++ src/hooks/useEffectiveSupply.ts
@@ -1,7 +1,7 @@
 import BigNumber from '@lib/bigNumber'
 import { useMemo } from 'react'
 import { PCT_BASE } from '../constants'
-import { ConfigType } from './constants'
+import { ConfigType } from '../types/app'
 
 export default function useEffectiveSupply(
   totalSupply: BigNumber,
