+++ src/hooks/useGardenTokenIcon.ts
@@ -1,6 +1,6 @@
 import { useConnectedGarden } from '@providers/ConnectedGarden'
 import { getGardenTokenIcon } from '../utils/token-utils'
-import { TokenType } from './constants'
+import { TokenType } from '../types/app'
 
 export default function useGardenTokenIcon(token: TokenType) {
   const connectedGarden = useConnectedGarden()
